stage 2 streaming split
